CloudRun2/requirements.txt
Flask==2.0.1
pytest==5.3.2; python_version > "3.0"
pytest==5.3.2; python_version < "3.0"
gunicorn==20.1.0
google-cloud-pubsub==1.7.0
google-cloud-bigquery==2.21.0 
pandas==1.3.3
google-cloud-bigquery-storage==2.9.0
pyarrow==6.0.0
google-cloud-storage==1.42.3